aster_time

Struct ClockSource

pub struct ClockSource { /* private fields */ }

ClockSource is an abstraction for hardware-assisted timing mechanisms.

A ClockSource can be created based on any counter that operates at a stable frequency. Users are able to measure time by retrieving Instant from this source.

Implementation

The ClockSource relies on obtaining the frequency of the counter and the method for reading the cycles in order to measure time.

The cycles here refer the counts of the base time counter.

Additionally, the ClockSource also holds a last record for an Instant and the corresponding cycles, which acts as a reference point for subsequent time retrieval. To prevent numerical overflow during the calculation of Instant, this last recorded instant must be periodically refreshed. The maximum interval for these updates must be determined at the time of the ClockSource initialization.

Examples

Suppose we have a counter called counter which have the frequency counter.freq, and the method to read its cycles called read_counter(). We can create a corresponding ClockSource and use it as follows:

// here we set the max_delay_secs = 10
let max_delay_secs = 10;
// create a clocksource named counter_clock
let counter_clock = ClockSource::new(counter.freq, max_delay_secs, Arc::new(read_counter));
// read an instant.
let instant = counter_clock.read_instant();

If using this ClockSource, you must ensure its internal instant will be updated at least once within a time interval of not more than `max_delay_secs.

Implementations

impl ClockSource

pub fn new( freq: u64, max_delay_secs: u64, read_cycles: Arc<dyn Fn() -> u64 + Sync + Send>, ) -> Self

Creates a new ClockSource instance.

This method requires basic information of the time counter, including the function for reading cycles, the frequency, and the maximum delay seconds to update this ClockSource. The ClockSource also calculates a reliable Coeff based on the counter’s frequency and the maximum delay seconds. This Coeff is used to convert the number of cycles into the duration of time that has passed for those cycles.

pub fn read_cycles(&self) -> u64

Reads current cycles of the ClockSource.

pub fn last_record(&self) -> (Instant, u64)

Returns the last instant and last cycles recorded in the ClockSource.

pub fn max_delay_secs(&self) -> u64

Returns the maximum delay seconds for updating of the ClockSource.

pub fn coeff(&self) -> &Coeff

Returns the reference to the generated cycles coeff of the ClockSource.

pub fn freq(&self) -> u64

Returns the frequency of the counter used in the ClockSource.